Rename yolo_inference to trt_inference.py
